исправление недочетов
