+++ js/filtration.js
@@ -1,13 +1,5 @@
 import {renderMiniatures} from './miniature.js';
-
-function debounce (callback, timeoutDelay = 500) {
-  let timeoutId;
-
-  return (...rest) => {
-    clearTimeout(timeoutId);
-    timeoutId = setTimeout(() => callback.apply(this, rest), timeoutDelay);
-  };
-}
+import {debounce} from './util.js';
 
 const filtrationForm = document.querySelector('.img-filters__form');
