+++ js/util.js
@@ -59,4 +59,13 @@ const showSuccess = () => {
   showMessage(successTamplate);
 };
 
-export {getRandomNumber, getCheckString, isEscapeKey, showError, showSuccess, getDataError};
+function debounce (callback, timeoutDelay = 500) {
+  let timeoutId;
+
+  return (...rest) => {
+    clearTimeout(timeoutId);
+    timeoutId = setTimeout(() => callback.apply(this, rest), timeoutDelay);
+  };
+}
+
+export {getRandomNumber, getCheckString, isEscapeKey, showError, showSuccess, getDataError, debounce};
